refactor: Controller response dto 제거
