+++ backend/src/main/java/kr/codesquad/issuetracker/application/IssueService.java
@@ -1,14 +1,12 @@
 package kr.codesquad.issuetracker.application;
 
-import java.util.Comparator;
 import java.util.List;
-import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import kr.codesquad.issuetracker.infrastructure.persistence.IssueRepository;
-import kr.codesquad.issuetracker.presentation.response.IssueSimpleResponse;
+import kr.codesquad.issuetracker.infrastructure.persistence.mapper.IssueSimpleMapper;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
@@ -18,9 +16,7 @@ public class IssueService {
 	private final IssueRepository issueRepository;
 
 	@Transactional(readOnly = true)
-	public List<IssueSimpleResponse> findAll() {
-		return issueRepository.findAll().stream()
-			.map(IssueSimpleResponse::from)
-			.collect(Collectors.toList());
+	public List<IssueSimpleMapper> findAll() {
+		return issueRepository.findAll();
 	}
 }
