+++ backend/src/main/java/kr/codesquad/issuetracker/presentation/IssueController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import kr.codesquad.issuetracker.application.IssueService;
-import kr.codesquad.issuetracker.presentation.response.IssueSimpleResponse;
+import kr.codesquad.issuetracker.infrastructure.persistence.mapper.IssueSimpleMapper;
 import lombok.RequiredArgsConstructor;
 
 @RequestMapping("/api/issues")
@@ -17,7 +17,7 @@ import lombok.RequiredArgsConstructor;
 public class IssueController {
 	private final IssueService issueService;
 	@GetMapping
-	public ResponseEntity<List<IssueSimpleResponse>> findAll() {
+	public ResponseEntity<List<IssueSimpleMapper>> findAll() {
 		return ResponseEntity.ok(issueService.findAll());
 	}
 }
